project/app/requirements.txt
fastapi
aiohttp
uvicorn
python-dotenv
openai
boto3
